Change apache vhost file from link to file
The naginator class defines an apache vhost
file that is intended to prevent the apache
module from purging our needed nagios vhost
files.

It was implemented as a self-referencial link.
This does not work starting in Puppet 3.3 (and
IMO should not work before that).

Basically, link targets create an implicity type
dependency. In Puppet > 3.3.x, this causes a
dependency loop for this resources.

This commit implements this resources as basic
    ensure => file
instead of a self-referential link.